blood-bank-api/requirements.txt
Flask
psycopg2-binary
argon2-cffi
flask-cors